requirements.txt
Django>=4.2,<5.0
djangorestframework
celery[redis]
gevent
psycopg2-binary
gunicorn
django-environ
requests